+++ test/raw.test.ts
@@ -1,0 +1,17 @@
+import { test } from "./utils.ts";
+
+Deno.test("Raw tag", async () => {
+  await test({
+    template: `
+    {{raw}} Hello world {{/raw}}
+    `,
+    expected: "Hello world",
+  });
+
+  await test({
+    template: `
+    {{raw}} Hello {{ world }} {{/raw}}
+    `,
+    expected: "Hello {{ world }}",
+  });
+});