+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.4.0] - 2023-06-14
 ### Added
 - Support for trims [#2].
   For example: `<h1> {{- title -}} </h1>`.
@@ -48,7 +48,7 @@ First version
 [#4]: https://github.com/oscarotero/vento/issues/4
 [#5]: https://github.com/oscarotero/vento/issues/5
 
-[Unreleased]: https://github.com/oscarotero/vento/compare/v0.3.1...HEAD
+[0.4.0]: https://github.com/oscarotero/vento/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/oscarotero/vento/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/oscarotero/vento/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/oscarotero/vento/compare/v0.1.0...v0.2.0